+++ src/RobotStudio.Domain/Articulated/ScaraLinkId.cs
@@ -1,0 +1,7 @@
+namespace RobotStudio.Domain.Articulated;
+
+public enum ScaraLinkId
+{
+    FirstLink,
+    SecondLink
+}